+++ MemoryAllocator/MemoryAllocator/MemoryManager.h
@@ -21,7 +21,7 @@ private:
 public:
 
 	PoolAllocator* CreatePoolAllocator(uint32_t sizeOfObject, uint32_t nrOfObjects);
-	StackAllocator* CreateStackAllocator(uint32_t size);
+	StackAllocator* CreateStackAllocator(uint64_t size);
 
 	//void ReleasePoolAllocator(PoolAllocator* object);
 	//void ReleaseStackAllocator(StackAllocator* object);
